grid list implemented in dialog
